Implement an Alt solution
Add `alt` as a dependency and implement a solution based off it to
complete the counter example.  This includes:
* A base `alt-app.js` file which initializes a single instance of alt
to use in the component.
* A `CounterActions` which doesn’t do much but describe the available
actions.  No input is sent so they both return null.
* A `CounterStore` which sets an initial state of the counter, and
binds the actions to the actual implementation which alters the counter
value.
* And wire this up together in the component via the `connectToStores`
“high order component”.

One thing that was a bit confusing was that the `connectToStores`
function had been moved to `alt-utils` in the latest version of Alt.
There wasn’t a lot of documentation about this change in the examples I
saw online.