Implement Redux solution
Add `redux` as a dependency and use it to implement the counter
example.  This includes:

* A set of `CounterActionTypes` which indicate the possible actions we
can take.
* The `CounterActions` which use those action types to declare
functions which return an object stating the type of action
(increment/decrement).
* The `CounterReducer` which sets up a default state (we could have
used a number for this but chose to use an object which includes the
count for a more complete example).  The `counter` function in this
reducer updates the state based on the action type.
* For a more complete example, an `index.js` reducer is supplied to
combine all reducers to a single root reducer.
* A generic `configureStore` implementation which used our combined
reducers.
* The wrapped App using the store, and the updated `App.js` which reads
in the state from the correct reducer, and delegates the correct action
to the `Counter.js` component.